OnCollisionEnter
added this method to the enemy movement. Now when the zombie hits the player, we can add some functionality for when that happens.  A prerequisite to this is that we must attach the Rigid body component to the zombie, despite the fact we disabled gravity and constrained it on all axis.